# spring/classes/openapi/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Abacus WEB Sample",
    "description": "Abacus WEB Sample Project",
    "license": {
      "name": "Apache 2.0",
      "url": "http://abacus.co.kr"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://192.168.2.142:8080/api/",
      "description": "dev"
    }
  ],
...
  "paths": {
...
    "/modem/updateModem": {
      "post": {
        "tags": [
          "서비스 등록 - 모듈 모델 등록"
        ],
        "summary": "모듈 모델 수정",
        "description": "모듈 모델 수정",
        "operationId": "updateModem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "modemCd": {
                    "type": "string",
                    "example": 1
                  },
                  "cmpCd": {
                    "type": "string",
                    "example": 1
                  },
                  "modemNm": {
                    "type": "string",
                    "example": "테스트 모듈 모델"
                  },
                  "modemManu": {
                    "type": "string",
                    "example": "admin"
                  },
                  "modemEmail": {
                    "type": "string",
                    "example": "test@test.com"
                  },
                  "anTypeList": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "10",
                      "10000000"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/modem/retrieveModem": {
      "post": {
        "tags": [
          "서비스 등록 - 모듈 모델 등록"
        ],
        "summary": "모듈 모델 상세 조회",
        "description": "모듈 모델 상세 조회",
        "operationId": "retrieveModem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "modemCd": {
                    "type": "string",
                    "example": 1
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComResponseModem"
                }
              }
            }
          }
        }
      }
    },
    "/modem/retrieveModemList": {
      "post": {
        "tags": [
          "서비스 등록 - 모듈 모델 등록"
        ],
        "summary": "모듈 모델 목록 조회",
        "description": "모듈 모델 목록 조회",
        "operationId": "retrieveModemList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "modemNm": {
                    "type": "string",
                    "example": ""
                  },
                  "cmpCd": {
                    "type": "string",
                    "example": ""
                  },
                  "currentPage": {
                    "type": "string",
                    "example": 1
                  },
                  "limit": {
                    "type": "string",
                    "example": 10
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComResponseModemList"
                }
              }
            }
          }
        }
      }
    },
    "/modem/downloadModemList": {
      "post": {
        "tags": [
          "서비스 등록 - 모듈 모델 등록"
        ],
        "summary": "모듈 모델 엑셀 다운로드",
        "description": "모듈 모델 엑셀 다운로드",
        "operationId": "downloadModemList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cmpCd": {
                    "type": "string",
                    "example": ""
                  },
                  "modemNm": {
                    "type": "string",
                    "example": ""
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/modem/deleteModem": {
      "post": {
        "tags": [
          "서비스 등록 - 모듈 모델 등록"
        ],
        "summary": "모듈 모델 삭제",
        "description": "모듈 모델 삭제",
        "operationId": "deleteModem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "modemCd": {
                    "type": "string",
                    "example": 1
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/modem/createModem": {
      "post": {
        "tags": [
          "서비스 등록 - 모듈 모델 등록"
        ],
        "summary": "모듈 모델 등록",
        "description": "모듈 모델 등록",
        "operationId": "createModem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cmpCd": {
                    "type": "string",
                    "example": 1
                  },
                  "modemNm": {
                    "type": "string",
                    "example": "테스트 모듈 모델"
                  },
                  "modemManu": {
                    "type": "string",
                    "example": "admin"
                  },
                  "modemEmail": {
                    "type": "string",
                    "example": "test@test.com"
                  },
                  "anTypeList": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "10",
                      "10000000"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/loglevel/applyLogLevel": {
      "post": {
        "tags": [
          "로그 조회 - 로그 레벨 조회"
        ],
        "summary": "로그 레벨 적용",
        "description": "로그 레벨 적용",
        "operationId": "applyLogLevel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogLevelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComResponseModem": {
        "type": "object",
        "properties": {
          "httpStatusCode": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detailMessage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/Modem"
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
      "Modem": {
        "required": [
          "anTypeList",
          "cmpCd",
          "modemCd",
          "modemEmail",
          "modemManu",
          "modemNm"
        ],
        "type": "object",
        "properties": {
          "useYn": {
            "type": "string"
          },
          "creDt": {
            "type": "string"
          },
          "creUsr": {
            "type": "string"
          },
          "updDt": {
            "type": "string"
          },
          "updUsr": {
            "type": "string"
          },
          "rolNm": {
            "type": "string"
          },
          "modemCd": {
            "pattern": "^[0-9]+$",
            "type": "string",
            "description": "모듈 모델 코드"
          },
          "modemNm": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "모듈 모델명"
          },
          "cmpCd": {
            "pattern": "^$|^[0-9]+$",
            "type": "string",
            "description": "회사 코드"
          },
          "cmpNm": {
            "type": "string",
            "description": "회사명"
          },
          "anTypeList": {
            "maxItems": 8,
            "minItems": 0,
            "type": "array",
            "description": "사용망 리스트 (화면표시를위한가공데이터)",
            "example": [
              10,
              10000000
            ],
            "items": {
              "type": "string",
              "description": "사용망 리스트 (화면표시를위한가공데이터)",
              "example": "[10,10000000]"
            }
          },
          "modemManu": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "description": "담당자",
            "example": "manager"
          },
          "modemEmail": {
            "maxLength": 50,
            "minLength": 0,
            "pattern": "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$",
            "type": "string",
            "description": "담당자 이메일",
            "example": "manager@abacus.com"
          },
          "anType": {
            "type": "string",
            "description": "사용망",
            "example": "10000010"
          }
        }
      },
      "ComResponseModemList": {
        "type": "object",
        "properties": {
          "httpStatusCode": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detailMessage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/ModemList"
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
      "ModemList": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Modem"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "LogLevelRequest": {
        "required": [
          "levelList",
          "serviceList"
        ],
        "type": "object",
        "properties": {
          "serviceList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "levelList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ttlMinutes": {
            "minimum": 10,
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}