# spring/generated-sources/openapi/src/main/java/com/openapi/generator/aicc/api/LoglevelApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.openapi.generator.aicc.api;

import com.openapi.generator.aicc.dto.ComResponseLogLevelList;
import com.openapi.generator.aicc.dto.LogLevelRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-19T16:30:08.398412+09:00[Asia/Seoul]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "로그 조회 - 로그 레벨 조회", description = "로그 레벨 조회")
public interface LoglevelApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_APPLY_LOG_LEVEL = "/loglevel/applyLogLevel";
    /**
     * POST /loglevel/applyLogLevel : 로그 레벨 적용
     * 로그 레벨 적용
     *
     * @param logLevelRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "applyLogLevel",
        summary = "로그 레벨 적용",
        description = "로그 레벨 적용",
        tags = { "로그 조회 - 로그 레벨 조회" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoglevelApi.PATH_APPLY_LOG_LEVEL,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> applyLogLevel(
        @Parameter(name = "LogLevelRequest", description = "", required = true) @Valid @RequestBody LogLevelRequest logLevelRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_LOG_LEVEL_LIST = "/loglevel/retrieveLogLevel";
    /**
     * POST /loglevel/retrieveLogLevel : 로그 레벨 목록 조회
     * 로그 레벨 목록 조회
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrieveLogLevelList",
        summary = "로그 레벨 목록 조회",
        description = "로그 레벨 목록 조회",
        tags = { "로그 조회 - 로그 레벨 조회" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponseLogLevelList.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoglevelApi.PATH_RETRIEVE_LOG_LEVEL_LIST,
        produces = { "application/json" }
    )
    default ResponseEntity<ComResponseLogLevelList> retrieveLogLevelList(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"list\" : [ { \"server\" : \"server\", \"currentTypes\" : [ \"currentTypes\", \"currentTypes\" ], \"applied\" : true }, { \"server\" : \"server\", \"currentTypes\" : [ \"currentTypes\", \"currentTypes\" ], \"applied\" : true } ], \"totalCount\" : 6 }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}