# spring/generated-sources/openapi/src/main/java/com/openapi/generator/aicc/api/ModemApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.openapi.generator.aicc.api;

import com.openapi.generator.aicc.dto.ComResponseModem;
import com.openapi.generator.aicc.dto.ComResponseModemList;
import com.openapi.generator.aicc.dto.CreateModemRequest;
import com.openapi.generator.aicc.dto.DownloadModemListRequest;
import com.openapi.generator.aicc.dto.RetrieveModemListRequest;
import com.openapi.generator.aicc.dto.RetrieveModemRequest;
import com.openapi.generator.aicc.dto.UpdateModemRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-19T16:30:08.398412+09:00[Asia/Seoul]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "서비스 등록 - 모듈 모델 등록", description = "서비스 등록 - 모듈 모델 등록")
public interface ModemApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_MODEM = "/modem/createModem";
    /**
     * POST /modem/createModem : 모듈 모델 등록
     * 모듈 모델 등록
     *
     * @param createModemRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "createModem",
        summary = "모듈 모델 등록",
        description = "모듈 모델 등록",
        tags = { "서비스 등록 - 모듈 모델 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ModemApi.PATH_CREATE_MODEM,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> createModem(
        @Parameter(name = "CreateModemRequest", description = "", required = true) @Valid @RequestBody CreateModemRequest createModemRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DELETE_MODEM = "/modem/deleteModem";
    /**
     * POST /modem/deleteModem : 모듈 모델 삭제
     * 모듈 모델 삭제
     *
     * @param retrieveModemRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deleteModem",
        summary = "모듈 모델 삭제",
        description = "모듈 모델 삭제",
        tags = { "서비스 등록 - 모듈 모델 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ModemApi.PATH_DELETE_MODEM,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> deleteModem(
        @Parameter(name = "RetrieveModemRequest", description = "", required = true) @Valid @RequestBody RetrieveModemRequest retrieveModemRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DOWNLOAD_MODEM_LIST = "/modem/downloadModemList";
    /**
     * POST /modem/downloadModemList : 모듈 모델 엑셀 다운로드
     * 모듈 모델 엑셀 다운로드
     *
     * @param downloadModemListRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "downloadModemList",
        summary = "모듈 모델 엑셀 다운로드",
        description = "모듈 모델 엑셀 다운로드",
        tags = { "서비스 등록 - 모듈 모델 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ModemApi.PATH_DOWNLOAD_MODEM_LIST,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> downloadModemList(
        @Parameter(name = "DownloadModemListRequest", description = "", required = true) @Valid @RequestBody DownloadModemListRequest downloadModemListRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_MODEM = "/modem/retrieveModem";
    /**
     * POST /modem/retrieveModem : 모듈 모델 상세 조회
     * 모듈 모델 상세 조회
     *
     * @param retrieveModemRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrieveModem",
        summary = "모듈 모델 상세 조회",
        description = "모듈 모델 상세 조회",
        tags = { "서비스 등록 - 모듈 모델 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponseModem.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ModemApi.PATH_RETRIEVE_MODEM,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponseModem> retrieveModem(
        @Parameter(name = "RetrieveModemRequest", description = "", required = true) @Valid @RequestBody RetrieveModemRequest retrieveModemRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"cmpCd\" : \"cmpCd\", \"cmpNm\" : \"cmpNm\", \"anTypeList\" : [ 10, 10000000 ], \"modemNm\" : \"modemNm\", \"modemEmail\" : \"manager@abacus.com\", \"modemCd\" : \"modemCd\", \"updUsr\" : \"updUsr\", \"creDt\" : \"creDt\", \"rolNm\" : \"rolNm\", \"modemManu\" : \"manager\", \"useYn\" : \"useYn\", \"creUsr\" : \"creUsr\", \"anType\" : \"10000010\", \"updDt\" : \"updDt\" }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_MODEM_LIST = "/modem/retrieveModemList";
    /**
     * POST /modem/retrieveModemList : 모듈 모델 목록 조회
     * 모듈 모델 목록 조회
     *
     * @param retrieveModemListRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrieveModemList",
        summary = "모듈 모델 목록 조회",
        description = "모듈 모델 목록 조회",
        tags = { "서비스 등록 - 모듈 모델 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponseModemList.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ModemApi.PATH_RETRIEVE_MODEM_LIST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponseModemList> retrieveModemList(
        @Parameter(name = "RetrieveModemListRequest", description = "", required = true) @Valid @RequestBody RetrieveModemListRequest retrieveModemListRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"list\" : [ { \"cmpCd\" : \"cmpCd\", \"cmpNm\" : \"cmpNm\", \"anTypeList\" : [ 10, 10000000 ], \"modemNm\" : \"modemNm\", \"modemEmail\" : \"manager@abacus.com\", \"modemCd\" : \"modemCd\", \"updUsr\" : \"updUsr\", \"creDt\" : \"creDt\", \"rolNm\" : \"rolNm\", \"modemManu\" : \"manager\", \"useYn\" : \"useYn\", \"creUsr\" : \"creUsr\", \"anType\" : \"10000010\", \"updDt\" : \"updDt\" }, { \"cmpCd\" : \"cmpCd\", \"cmpNm\" : \"cmpNm\", \"anTypeList\" : [ 10, 10000000 ], \"modemNm\" : \"modemNm\", \"modemEmail\" : \"manager@abacus.com\", \"modemCd\" : \"modemCd\", \"updUsr\" : \"updUsr\", \"creDt\" : \"creDt\", \"rolNm\" : \"rolNm\", \"modemManu\" : \"manager\", \"useYn\" : \"useYn\", \"creUsr\" : \"creUsr\", \"anType\" : \"10000010\", \"updDt\" : \"updDt\" } ], \"totalCount\" : 6 }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPDATE_MODEM = "/modem/updateModem";
    /**
     * POST /modem/updateModem : 모듈 모델 수정
     * 모듈 모델 수정
     *
     * @param updateModemRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateModem",
        summary = "모듈 모델 수정",
        description = "모듈 모델 수정",
        tags = { "서비스 등록 - 모듈 모델 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ModemApi.PATH_UPDATE_MODEM,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> updateModem(
        @Parameter(name = "UpdateModemRequest", description = "", required = true) @Valid @RequestBody UpdateModemRequest updateModemRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}